Added tests for the monetization settings. This includes tests for setting transfers from the customizer over to the settings panel. The transfer of customizer settings was not checking if every field was empty and made the assumption that there was no scenario where settings could be saved in both the customizer and the settings panel. I believe this scenario can arise if people switch between versions and tweaking this makes the function more robust and ensures there is no chance of leaving unused variables set in the theme_mod. This applies as much for the visual settings as it does for the messaging settings. Removed the previous tests file which has been incorporated in the monetization testing

+++ includes/settings/functions.php
@@ -970,10 +970,6 @@ function transfer_customizer_message_settings() {
 
 	$existing_options = get_option( 'coil_messaging_settings_group' );
 
-	if ( ! empty( $existing_options ) ) {
-		return;
-	}
-
 	$messaging_settings = [];
 
 	$coil_partial_gating_message     = 'coil_partial_gating_message';
@@ -984,12 +980,26 @@ function transfer_customizer_message_settings() {
 	$coil_learn_more_button_text     = 'coil_learn_more_button_text';
 	$coil_learn_more_button_link     = 'coil_learn_more_button_link';
 
+	$customizer_empty = (
+		get_theme_mod( $coil_partial_gating_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_unsupported_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_verifying_status_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_unable_to_verify_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_voluntary_donation_message, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_learn_more_button_text, 'null' ) !== 'null'
+		&& get_theme_mod( $coil_learn_more_button_link, 'null' ) !== 'null'
+	);
+
+	if ( $customizer_empty ) {
+		return;
+	}
+
 	// Using 'null' for comparrison becasue custom messages that were deleted remain in the database with the value false, but still need to be removed.
 	if ( get_theme_mod( $coil_partial_gating_message, 'null' ) !== 'null' ) {
 		$messaging_settings['coil_partially_gated_content_message'] = get_theme_mod( $coil_partial_gating_message );
 		remove_theme_mod( $coil_partial_gating_message );
 	}
-	if ( get_theme_mod( $coil_unsupported_message , 'null' ) !== 'null' ) {
+	if ( get_theme_mod( $coil_unsupported_message, 'null' ) !== 'null' ) {
 		$messaging_settings['coil_fully_gated_content_message'] = get_theme_mod( $coil_unsupported_message );
 		remove_theme_mod( $coil_unsupported_message );
 	}
@@ -1019,7 +1029,11 @@ function transfer_customizer_message_settings() {
 		remove_theme_mod( $coil_learn_more_button_link );
 	}
 
-	update_option( 'coil_messaging_settings_group', $messaging_settings );
+	if ( false !== $existing_options ) {
+		update_option( 'coil_messaging_settings_group', array_merge( $existing_options, $messaging_settings ) );
+	} else {
+		update_option( 'coil_messaging_settings_group', $messaging_settings );
+	}
 
 }
 
@@ -1031,23 +1045,26 @@ function transfer_customizer_monetization_settings() {
 
 	// If the setting has already been saved or transferred then simply return
 	// Using 'null' for comparrison becasue if the padlock and support creator messages were unselected they were stored in the database with the value false, but still need to be transferred.
-	if ( get_option( 'coil_monetization_settings_group' ) || ( ! get_option( 'coil_content_settings_posts_group' ) && 'null' === get_theme_mod( 'coil_title_padlock', 'null' ) && 'null' === get_theme_mod( 'coil_show_donation_bar', 'null' ) ) ) {
+	if ( get_option( 'coil_monetization_settings_group' ) && ! get_option( 'coil_content_settings_posts_group' ) && 'null' === get_theme_mod( 'coil_title_padlock', 'null' ) && 'null' === get_theme_mod( 'coil_show_donation_bar', 'null' ) ) {
 		return;
 	}
 
-	$coil_title_padlock     = 'coil_title_padlock';
-	$coil_show_donation_bar = 'coil_show_donation_bar';
+	$coil_title_padlock        = 'coil_title_padlock';
+	$coil_show_donation_bar    = 'coil_show_donation_bar';
+	$new_monetization_settings = [];
 
-	$new_monetization_settings = [
-		'coil_title_padlock'     => get_theme_mod( $coil_title_padlock, true ),
-		'coil_show_donation_bar' => get_theme_mod( $coil_show_donation_bar, true ),
-	];
+	if ( get_theme_mod( $coil_title_padlock, 'null' ) !== 'null' ) {
+		$new_monetization_settings['coil_title_padlock'] = get_theme_mod( $coil_title_padlock, true );
+		remove_theme_mod( $coil_title_padlock );
+	}
 
-	remove_theme_mod( $coil_title_padlock );
-	remove_theme_mod( $coil_show_donation_bar );
+	if ( get_theme_mod( $coil_show_donation_bar, 'null' ) !== 'null' ) {
+		$new_monetization_settings['coil_show_donation_bar'] = get_theme_mod( $coil_show_donation_bar, true );
+		remove_theme_mod( $coil_show_donation_bar );
+	}
 
 	// Before moving post gating settings across check that post types and gating types are valid.
-	$supported_post_types = Coil\get_supported_post_types( 'names' );
+	$supported_post_types     = Coil\get_supported_post_types( 'names' );
 	$supported_gating_options = array_keys( Gating\get_monetization_setting_types() );
 
 	if ( ! empty( $supported_post_types ) && ! empty( $previous_gating_options ) ) {
